+++ front/src/views/bill/bill/summary/product/index.tsx
@@ -5,6 +5,7 @@ import { useI18n } from 'vue-i18n';
 import Amount from '../../components/amount';
 import Search from '../../components/search';
 
+import changeCurrency from '../change-currency';
 import useColumns from '@/views/resource/resource-manage/hooks/use-columns';
 import { useTable } from '@/hooks/useTable/useTable';
 import { reqBillsMainAccountSummarySum } from '@/api/bill';
@@ -35,7 +36,8 @@ export default defineComponent({
       renderOperation,
     } = useProductHandler();
 
-    const { columns, handleChangeCurrencyChecked } = useColumns(columnName);
+    const { handleChangeCurrencyChecked, getColElement } = changeCurrency();
+    const { columns } = useColumns(columnName, false, '', { getColElement });
     const { CommonTable, getListData, clearFilter, filter } = useTable({
       searchOptions: { disabled: true },
       tableOptions: {
