+++ front/src/views/bill/bill/summary/sub/index.tsx
@@ -8,6 +8,7 @@ import Amount from '../../components/amount';
 import { useI18n } from 'vue-i18n';
 import { useTable } from '@/hooks/useTable/useTable';
 import useColumns from '@/views/resource/resource-manage/hooks/use-columns';
+import changeCurrency from '../change-currency';
 import {
   exportBillsMainAccountSummary,
   reqBillsMainAccountSummaryList,
@@ -31,7 +32,8 @@ export default defineComponent({
     const { useSubHandler } = pluginHandler;
     const { mountedCallback } = useSubHandler();
 
-    const { columns, handleChangeCurrencyChecked } = useColumns('billsMainAccountSummary');
+    const { handleChangeCurrencyChecked, getColElement } = changeCurrency();
+    const { columns } = useColumns('billsMainAccountSummary', false, '', { getColElement });
     const { CommonTable, getListData, clearFilter, filter } = useTable({
       searchOptions: { disabled: true },
       tableOptions: {
