+++ front/src/views/resource/resource-manage/hooks/use-columns.tsx
@@ -32,11 +32,11 @@ import {
   SCHEDULER_MAP,
   TRANSPORT_LAYER_LIST,
 } from '@/constants/clb';
-import { formatBillCost, getInstVip, formatBillRatio, formatBillRatioClass, formatBillSymbol } from '@/utils';
-import { Spinner, ArrowsRight } from 'bkui-vue/lib/icon';
+import { formatBillCost, getInstVip, formatBillRatio, formatBillRatioClass } from '@/utils';
+import { Spinner } from 'bkui-vue/lib/icon';
 import { APPLICATION_STATUS_MAP, APPLICATION_TYPE_MAP } from '@/views/service/apply-list/constants';
 import dayjs from 'dayjs';
-import { BILLS_ROOT_ACCOUNT_SUMMARY_STATE_MAP, BILL_TYPE__MAP_HW, CURRENCY_MAP, CURRENCY_ALIAS_MAP } from '@/constants';
+import { BILLS_ROOT_ACCOUNT_SUMMARY_STATE_MAP, BILL_TYPE__MAP_HW, CURRENCY_MAP } from '@/constants';
 import { BILL_VENDORS_MAP, BILL_SITE_TYPES_MAP } from '@/views/bill/account/account-manage/constants';
 import CopyToClipboard from '@/components/copy-to-clipboard/index.vue';
 
@@ -54,8 +54,8 @@ interface LinkFieldOptions {
   sort?: boolean; // 是否支持排序
 }
 
-export default (type: string, isSimpleShow = false, vendor?: string) => {
-  const changeCurrencyChecked = ref(false);
+export default (type: string, isSimpleShow = false, vendor?: string, options?: any) => {
+  const { getColElement } = options;
   const router = useRouter();
   const route = useRoute();
   const accountStore = useAccountStore();
@@ -124,19 +124,6 @@ export default (type: string, isSimpleShow = false, vendor?: string) => {
     };
   };
 
-  const getColElement = (money: string, converted: string, currency: string) => {
-    const normalData = formatBillSymbol(money, currency);
-    if (!changeCurrencyChecked.value || currency !== CURRENCY_ALIAS_MAP.USD) {
-      return normalData;
-    }
-    return (
-      <div class={'current-currency'}>
-        <span class={'dollar'}>{normalData}</span>
-        <ArrowsRight class={'arrow-right'} />
-        <span> {formatBillSymbol(converted, CURRENCY_ALIAS_MAP.CNY)} </span>
-      </div>
-    );
-  };
   /**
    * todo: 更换实现方式, 取消使用stopPropagation
    * 自定义 render field 的 push 导航
@@ -2008,7 +1995,7 @@ export default (type: string, isSimpleShow = false, vendor?: string) => {
       isDefaultShow: true,
       render: ({ data }: any) => {
         const { current_month_cost_synced, currency, current_month_rmb_cost_synced } = data;
-        return getColElement(current_month_cost_synced, current_month_rmb_cost_synced, currency);
+        return getColElement?.(current_month_cost_synced, current_month_rmb_cost_synced, currency);
       },
     },
     {
@@ -2030,7 +2017,7 @@ export default (type: string, isSimpleShow = false, vendor?: string) => {
           field: 'current_month_cost',
           render: ({ data }: any) => {
             const { current_month_cost, currency, current_month_rmb_cost } = data;
-            return getColElement(current_month_cost, current_month_rmb_cost, currency);
+            return getColElement?.(current_month_cost, current_month_rmb_cost, currency);
           },
         },
         {
@@ -2039,7 +2026,7 @@ export default (type: string, isSimpleShow = false, vendor?: string) => {
           field: 'last_month_cost_synced',
           render: ({ data }: any) => {
             const { last_month_cost_synced, currency, last_month_rmb_cost_synced } = data;
-            return getColElement(last_month_cost_synced, last_month_rmb_cost_synced, currency);
+            return getColElement?.(last_month_cost_synced, last_month_rmb_cost_synced, currency);
           },
         },
         {
@@ -2060,7 +2047,7 @@ export default (type: string, isSimpleShow = false, vendor?: string) => {
       isDefaultShow: true,
       render: ({ data }: any) => {
         const { adjustment_cost, currency, adjustment_rmb_cost } = data;
-        return getColElement(adjustment_cost, adjustment_rmb_cost, currency);
+        return getColElement?.(adjustment_cost, adjustment_rmb_cost, currency);
       },
     },
   ];
@@ -2110,7 +2097,7 @@ export default (type: string, isSimpleShow = false, vendor?: string) => {
       isDefaultShow: true,
       render: ({ data }: any) => {
         const { current_month_cost_synced, currency, current_month_rmb_cost_synced } = data;
-        return getColElement(current_month_cost_synced, current_month_rmb_cost_synced, currency);
+        return getColElement?.(current_month_cost_synced, current_month_rmb_cost_synced, currency);
       },
     },
     {
@@ -2119,7 +2106,7 @@ export default (type: string, isSimpleShow = false, vendor?: string) => {
       isDefaultShow: true,
       render: ({ data }: any) => {
         const { current_month_cost, currency, current_month_rmb_cost } = data;
-        return getColElement(current_month_cost, current_month_rmb_cost, currency);
+        return getColElement?.(current_month_cost, current_month_rmb_cost, currency);
       },
     },
   ];
@@ -2622,14 +2609,9 @@ export default (type: string, isSimpleShow = false, vendor?: string) => {
 
   const settings = generateColumnsSettings(columns);
 
-  const handleChangeCurrencyChecked = (val: boolean) => {
-    changeCurrencyChecked.value = val;
-  };
-
   return {
     columns,
     settings,
     generateColumnsSettings,
-    handleChangeCurrencyChecked,
   };
 };
